One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/pl-merch/pl-merch.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { PlMerchComponent } from './pl-merch.component';

describe('Component: PlMerch', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [PlMerchComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([PlMerchComponent],
      (component: PlMerchComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(PlMerchComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(PlMerchComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-pl-merch></app-pl-merch>
  `,
  directives: [PlMerchComponent]
})
class PlMerchComponentTestController {
}

